requirements.txt
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
filelock==3.20.3
fsspec==2026.1.0
idna==3.11
Jinja2==3.1.6
llvmlite==0.46.0
MarkupSafe==3.0.3
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.6.1
numba==0.63.1
numpy==2.3.5
openai-whisper==20250625
pycparser==3.0
pynput==1.8.1
pyperclip==1.11.0
regex==2026.1.15
requests==2.32.5
scipy==1.17.0
setuptools==80.10.2
six==1.17.0
sounddevice==0.5.5
sympy==1.14.0
tiktoken==0.12.0
torch==2.10.0
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.6.3
